Dev to main - about 12 months of on and off work (#1)
Merging the `dev` branch into `main`, the first of such merges was always planned for when functions were implemented, which they are now.

This represents 221 commits over 12 months, of which the last 4 months represent a surge in effort to get working language features.

The result of this effort includes:

- Pratt-style parser
- Abstract Syntax Tree (AST)
- High-level Intermediate Representation (HIR) using arenas for memory efficiency
- Type checker
- Bytecode generation
- Virtual machine bytecode executor
- Test suite for End-to-End tests

The scope of language features includes:
- Builtin types: Int, Float, Bool, String
- Type checker using bidirectional algorithm
- Variables
- Arithmetic operations (`+`, `-`, `*`, `/`)
- Unary operations (`-`, `!`)
- Temporary "to string" unary operator `~` useful for printing Int/Float/Bool
- Builtin functions (just `print` for now)
- Functions and calls

This merge will be a squash & merge. I don't intend to collect fine-grained git history at this time, but will do so when the language starts to approach a "release".