Assorted changes from dev branch (#8)
* Create new crates for Language Server

lsp-server: implementation of the actual server using tower-lsp
lsp-diagnostic: implementation of a diagnostic type to use in the rest of the program. May remove or rework this later to ensure LSPDiagnostic is just one output of the core diagnostics (String would be another output, for example, to print diagnostics when used from CLI)

* Add ability to codegen for individual functions

Currently useful for tests, but later may be used for incremental compilation and/or hot reloading

* Removed LSPDiagnostic from the CLI crate

Moved to its own crate

* Updated dependencies after adding tower-lsp dependency

* Removed serde dependency from CLI

After moving LSPDiagnostic to its own crate, the CLI doesn't need serde anymore (currently)

* Add note in the README about running with MIRI

Currently is more of a reminder to not have to look it up every time. Not ready to consistently test with MIRI

* Refactor/simplification for hir::Context

Context now has ownership of its Interner rather than a mutable borrow. Makes everything simpler, removes lifetime annotations, etc.

* Small adjustment to the spelling of `Unit` (`()`) in the examples

* Start a crate for the MIR representation

Similar to rustc, the MIR will be a Control Flow Graph. The idea eventually is to generate the Cranelift code (CLIF IR) from this MIR.

Other codegen outputs may not use this representation, for example, JavaScript may be generated from the HIR since JS is higher-level CLIF IR