Delete top-level lsp folder
The original intention I think was to have a Node server as the Language Server that would call into the compiler via FFI. This top-level folder was supposed to house both the server and the VS Code client.

However, now I feel it would be better to implement the language server in Rust so that it would be available as a library and compiled into the same binary as the rest of the compiler. This is removed in favor of crates inside the normal crates folder and a separate project in clients/vscode for the client.