[Breaking] Remove VM and all related crates
Originally, this language was going to be compiled to a custom bytecode format and executed by a custom virtual machine. This direction has changed, and the primary backend for this compiler will be to Cranelift IR (CLIF).

- The level of effort to implement language features in the bytecode and VM was substantial
- Relatively few resources for designing a VM for statically typed languages (many resources for Lua-like, JS-like, etc.). Values having different sizes on the stack was difficult to find (even the JVM predominately uses "everything is a pointer" on the stack).
- VM used a lot of unsafe code that was likely incorrect. Several "invalid free" bugs during development that were difficult to resolve.

In the future, if any additional backends are added for bytecode VM, it should be to existing VM formats such as the JVM or CLR (.NET).