Adds an experimental JavaScript code generation back-end
Very much not feature-complete, but contains the ground work for generating JavaScript code from the compiler.

The JS Codegen works from the HIR, not the MIR. This should still be the case when the MIR is completed because the HIR is a better representation for JS (higher level), whereas the MIR is lower level and is a better representation for CLIF generation.

JavaScript has similar scoping rules, one of the few differences will be that variable shadowing is not allowed in the same scope so we'll need to generate a unique variable name. Another difference is needing to define-then-mutate a variable because blocks are not expressions in JS.