Rework MIR construction to handle nested scopes better
A HIR block without control flow, i.e. just a regular block that has a single predecessor and then falls through back to the original scope can be constructed within the same MIR BasicBlock.

Set up machinery to track Scope during MIR construction. Experiment with a queue to process blocks, but may remove that in the future depending on if it's helpful for constructing SwitchInt terminators.