Implement mutability
Variables are able to be defined as mutable in the let binding. Currently, this allows for reassignment.

An explicit check for reassignment of non-mutable variables is not implemented yet, but does mostly "work" by virtue of the literal type being inferred. The type of `let a = 10` is inferred as `10`, so reassigning to a different value is a type error.